# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ArP - Artwork Provenance API
  description: |
    REST API for managing artwork provenance data using Semantic Web technologies.
    
    This API provides endpoints for:
    - Managing artworks (CRUD operations)
    - Tracking provenance chains (ownership history)
    - Executing SPARQL queries against the RDF knowledge base
    - Full-text search with faceted filtering
    - Artwork recommendations
    - QR code generation for artwork identification
    - **External SPARQL Integration**: Query DBpedia, Wikidata, and Getty AAT for enriched artwork metadata
  version: 1.0.0
  contact:
    name: ArP Development Team
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api
    description: Local development server
  - url: https://api.arp.example.org/api
    description: Production server
...
paths:
...
  /external/dbpedia/artwork/{resource}:
    get:
      tags: [external]
      summary: Query DBpedia for artwork information
      description: Fetch artwork metadata from DBpedia including abstract, thumbnail, artist, and museum information.
      operationId: getDbpediaArtwork
      parameters:
        - name: resource
          in: path
          required: true
          description: "DBpedia resource name or full URI (e.g., 'Mona_Lisa')"
          schema:
            type: string
      responses:
        '200':
          description: DBpedia data retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExternalQueryResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /external/dbpedia/artist/{resource}:
    get:
      tags: [external]
      summary: Query DBpedia for artist information
      description: Fetch artist biographical data from DBpedia including birth/death dates, nationality, and art movement.
      operationId: getDbpediaArtist
      parameters:
        - name: resource
          in: path
          required: true
          description: "DBpedia resource name or full URI (e.g., 'Leonardo_da_Vinci')"
          schema:
            type: string
      responses:
        '200':
          description: DBpedia artist data retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExternalQueryResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /external/wikidata/artwork/{entity}:
    get:
      tags: [external]
      summary: Query Wikidata for artwork information
      description: Fetch artwork metadata from Wikidata including image, inception date, materials, and location.
      operationId: getWikidataArtwork
      parameters:
        - name: entity
          in: path
          required: true
          description: "Wikidata Q-ID or full URI (e.g., 'Q12418')"
          schema:
            type: string
      responses:
        '200':
          description: Wikidata data retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExternalQueryResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /external/wikidata/artist/{entity}:
    get:
      tags: [external]
      summary: Query Wikidata for artist information
      description: Fetch artist biographical data from Wikidata including birth/death dates, nationality, and occupation.
      operationId: getWikidataArtist
      parameters:
        - name: entity
          in: path
          required: true
          description: "Wikidata Q-ID or full URI (e.g., 'Q762' for Leonardo da Vinci)"
          schema:
            type: string
      responses:
        '200':
          description: Wikidata artist data retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExternalQueryResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /external/getty/{aat_id}:
    get:
      tags: [external]
      summary: Query Getty AAT for term information
      description: Fetch term information from Getty Art & Architecture Thesaurus including preferred label and scope note.
      operationId: getGettyTerm
      parameters:
        - name: aat_id
          in: path
          required: true
          description: "Getty AAT ID (e.g., '300015050' for oil paint)"
          schema:
            type: string
      responses:
        '200':
          description: Getty AAT term retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExternalQueryResponse'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
...
  schemas:
...
    ExternalQueryResponse:
      type: object
      description: Generic response for external SPARQL queries
      properties:
        success:
          type: boolean
        data:
          type: object
          additionalProperties: true
        error:
          type: string
        source:
          type: string
          enum: [dbpedia, wikidata, getty]
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Error code
              example: "NOT_FOUND"
            message:
              type: string
              description: Human-readable error message
              example: "Artwork not found"
            details:
              type: object
              additionalProperties: true
              description: Additional error details
  responses:
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "INTERNAL_ERROR"
              message: "An unexpected error occurred"